Extract viewChangelog method to Utils.kt
